Breaking message module into own file
